menu option to set primary attachment for item
Tool -> Manage Attachments has a menu item to set an attachment
as a primary file by adding a relation to the parent item with
attachment's itemID as the object.
Primary attachment will be opened next time on click on the item.
If the primary attachment is moved to trash, it is ignored.
When a primary attachment is deleted or moved to another parent,
the relation is deleted.
getAttachments() function sorts attachments, and if there is
a primary attachment, it is moved to the top of the list.

Fixes: zotero#355
Fixes: zotero#1715