annotations showing in itemTree
- Annotations are displayed in itemTree under their file attachments
on a third level. The annotation spans the entire row.
- The title is constructed on the go and it attempts to be
most descriptive collection of information about the annotation:
	+ Start with a comment surrounded by " for visibility
	+ If there is still space (< 150 characters, might need to tweak the exact
		count), add annotation text.
	+ If there is still space, add tags.
	+ If the annotation type is image or link, prepend annotation type to the beginning.
	+ Add "|" between these components for separation.
- Search matches the actual attachment instead of it's parent file.
- Can create child notes from annotations of the same item or
standalone notes from annotations across different items from
the context menu.
- When an annotation (or multiple annotations) are selected, the
pane header is hidden and the only collapsible section in the itemPane
is the annotation box showing the preview of the selected annotation(s)