annotation deletion from itemTree
- removed "Move to trash" from context menu
- special treatment for annotations to always prompt
to erase the item regardless of what collectionTree row
is selected (e.g. if a collection is selected, we
still want one to be able to delete the annotation).
- this only applies if all selected items are annotations.
If multiple items are selected, some annotations and
some not, do nothing. This is until the trash is
ready.
- for now, the annotation is erased. But we'll want
to send them to trash, from where one should be able
to finally erase them or to restore.